requirements.txt
fastapi==0.116.1
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.5
pdfplumber==0.11.7
python-docx==1.2.0
python-dotenv==1.1.1
requests==2.32.4
uvicorn==0.35.0
websockets==15.0.1
# pinecone==7.3.0
tiktoken==0.9.0

